very_average_golfers/requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
boto3==1.9.200
botocore==1.12.200
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
dj-database-url==0.5.0
Django==2.2.1
django-appconf==1.0.3
django-braces==1.13.0
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-imagekit==4.0.2
django-js-asset==1.2.2
django-storages==1.7.1
django-tables2==2.0.5
djangoajax==3.1
docutils==0.14
easy-thumbnails==2.6
fabric==2.4.0
gunicorn==19.9.0
idna==2.8
invoke==1.2.0
jmespath==0.9.4
paramiko==2.4.2
pilkit==2.0
Pillow==5.4.1
psycopg2==2.8.3
pyasn1==0.4.5
pycparser==2.19
Pygments==2.4.0
PyNaCl==1.3.0
python-dateutil==2.8.0
pytz==2018.9
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
sorl-thumbnail==12.5.0
South==1.0.2
sqlparse==0.3.0
Unipath==1.1
urllib3==1.25.2
whitenoise==4.1.3
